+++ NWN.Anvil/src/main/API/Events/Game/CreatureEvents/CreatureEvents.OnConversation.cs
@@ -10,14 +10,20 @@ namespace Anvil.API.Events
   public static partial class CreatureEvents
   {
     /// <summary>
-    /// Triggered when <see cref="NwCreature"/> begins a conversation.
+    /// Called when this creature starts a conversation, or hears a message they are listening for.
     /// </summary>
     [GameEvent(EventScriptType.CreatureOnDialogue)]
     public sealed class OnConversation : IEvent
     {
       /// <summary>
+      /// Gets the creature attached to this conversation.
+      /// </summary>
+      public NwCreature Creature { get; } = NWScript.OBJECT_SELF.ToNwObject<NwCreature>()!;
+
+      /// <summary>
       /// Gets the <see cref="NwCreature"/> currently speaking.
       /// </summary>
+      [Obsolete("Use the Creature property instead.")]
       public NwCreature? CurrentSpeaker { get; } = NWScript.OBJECT_SELF.ToNwObject<NwCreature>();
 
       /// <summary>
@@ -30,7 +36,17 @@ namespace Anvil.API.Events
       /// </summary>
       public NwPlayer? PlayerSpeaker { get; } = NWScript.GetPCSpeaker().ToNwPlayer();
 
-      NwObject? IEvent.Context => CurrentSpeaker;
+      /// <summary>
+      /// Gets the listen pattern that matched the message sent to this creature.
+      /// </summary>
+      public int ListenPattern { get; } = NWScript.GetListenPatternNumber();
+
+      /// <summary>
+      /// Gets the associate command that matched the message sent to this creature.
+      /// </summary>
+      public AssociateCommand AssociateCommand => (AssociateCommand)ListenPattern;
+
+      NwObject IEvent.Context => Creature;
 
       public static void Signal(NwCreature creature)
       {